Web_App/requirements.txt
pandas==1.5.0
numpy==1.23.4
requests==2.28.1
streamlit==1.15.1
streamlit-folium==0.7.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
folium==0.13.0
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
geopy==2.2.0
branca==0.6.0
plotly==5.11.0
plotly-express==0.4.1
